airflow/requirements.txt
apache-airflow[amazon]==2.10.3
pandas==2.2.3
python-dotenv==1.0.1
scipy==1.14.1